Move embed implementation to the command itself
Reason: /active/impls/ is for interactable commands, and relaxation
vault doesn't have enough data to warrant having two embeds and whatnot
