+++ bathbot/src/active/impls/mod.rs
@@ -60,7 +60,6 @@ mod profile;
 mod ranking;
 mod ranking_countries;
 mod recent_list;
-pub(crate) mod relax;
 mod render;
 mod simulate;
 mod single_score;
